chore: Refactor diff and metadata module structure

+++ crates/wamex-cli/src/diff/symbols_map/graph_utils.rs
@@ -40,7 +40,7 @@ pub enum Child {
     Function(InputFuncId),
     DataSymbol {
         id: (DataSegmentId, DataSymbolId),
-        offset: u32,
+        offset: i64,
     },
     OtherReloc(OtherReloc),
 }
